Updated examples to new rwlock implementation
